requirements.txt
matplotlib==3.5.0
Pillow>=9.0.1
PyOpenGL
PyOpenGL_accelerate
PyYAML>=6.0
yacs>=0.1.8
scikit-image==0.19.1
termcolor
tqdm==4.62.3
loguru==0.5.3
trimesh==3.9.35
flatten_dict==0.4.2
ipykernel==5.3.4
ipywidgets==7.6.5
jpeg4py
shapely==1.7.1
vedo==2020.4.2
rtree==0.9.7
pytorch_lightning==1.2.5
PyMCubes
kornia
chumpy
tinyobjloader==2.0.0rc7
cython==0.29.20
opencv-python
opencv_contrib_python
scikit-learn
simple-romp==1.0.4
protobuf==3.20.0
pymeshlab
dataclasses>=0.6
numpy==1.22.4
